// src/game/world/generated/areaPieces.ts
export const AREA_SHEET = {
  imagePath: 'assets/frlg-areas.png',
  columns: 20,
  rows: 25,
} as const;
// ...
export type AreaPieceName = keyof typeof AREA_PIECES;